requirements.txt
pandas
openbb_terminal.sdk
numpy
matplotlib
scipy
sklearn
plotly
PIL